pqm4-specific changes and new files

* pqm4-specific changes and new files

* Speed up M4 arithmetic for p248

* Change some data types to size_t

* Change randombytes parameter to size_t


Approved-by: Isaac.Canales
Approved-by: Jorge Chavez

+++ include/rng.h
@@ -23,6 +23,6 @@ void randombytes_init(unsigned char *entropy_input,
  * @param[in] xlen Number of random bytes to be generated
  * @return int 0 on success, -1 otherwise
  */
-int randombytes(unsigned char *x, unsigned long long xlen);
+int randombytes(unsigned char *x, size_t xlen);
 
 #endif /* rng_h */
