+++ test/bench.c
@@ -85,7 +85,7 @@ static int bench_sig(int runs, int csv) {
     unsigned char *sk  = calloc(CRYPTO_SECRETKEYBYTES, 1);
     unsigned char *sig = calloc(CRYPTO_BYTES + m_len, 1);
     unsigned char *m   = calloc(m_len, 1);
-    unsigned long long len = CRYPTO_BYTES;
+    size_t len = CRYPTO_BYTES;
 
     if (csv) {
         printf("%s,", CRYPTO_ALGNAME);
